requirements.txt
black==19.10b0
Django==2.2.17
djangorestframework==3.12.2
ipdb==0.13.4
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-pythonpath==0.7.3
